requirements.txt
streamlit
googlesearch-python
pandas